Создал public class DevelopmentConfig для запуска с уже зарегистрированным пользователем
