+++ src/main/java/axsoup/AxsoupCloudApplication.java
@@ -17,22 +17,22 @@ public class AxsoupCloudApplication {
 		SpringApplication.run(AxsoupCloudApplication.class, args);
 	}
 
-	@Bean
-	public CommandLineRunner dataLoader(IngredientRepository repo) {
-		return new CommandLineRunner() {
-			@Override
-			public void run(String... args) throws Exception {
-				repo.save(new Ingredient("FLTO", "Flour Tortilla", Type.WRAP));
-				repo.save(new Ingredient("COTO", "Corn Tortilla", Type.WRAP));
-				repo.save(new Ingredient("GRBF", "Ground Beef", Type.PROTEIN));
-				repo.save(new Ingredient("CARN", "Carnitas", Type.PROTEIN));
-				repo.save(new Ingredient("TMTO", "Diced Tomatoes", Type.VEGGIES));
-				repo.save(new Ingredient("LETC", "Lettuce", Type.VEGGIES));
-				repo.save(new Ingredient("CHED", "Cheddar", Type.CHEESE));
-				repo.save(new Ingredient("JACK", "Monterrey Jack", Type.CHEESE));
-				repo.save(new Ingredient("SLSA", "Salsa", Type.SAUCE));
-				repo.save(new Ingredient("SRCR", "Sour Cream", Type.SAUCE));
-			}
-		};
-	}
+//	@Bean
+//	public CommandLineRunner dataLoader(IngredientRepository repo) {
+//		return new CommandLineRunner() {
+//			@Override
+//			public void run(String... args) throws Exception {
+//				repo.save(new Ingredient("FLTO", "Flour Tortilla", Type.WRAP));
+//				repo.save(new Ingredient("COTO", "Corn Tortilla", Type.WRAP));
+//				repo.save(new Ingredient("GRBF", "Ground Beef", Type.PROTEIN));
+//				repo.save(new Ingredient("CARN", "Carnitas", Type.PROTEIN));
+//				repo.save(new Ingredient("TMTO", "Diced Tomatoes", Type.VEGGIES));
+//				repo.save(new Ingredient("LETC", "Lettuce", Type.VEGGIES));
+//				repo.save(new Ingredient("CHED", "Cheddar", Type.CHEESE));
+//				repo.save(new Ingredient("JACK", "Monterrey Jack", Type.CHEESE));
+//				repo.save(new Ingredient("SLSA", "Salsa", Type.SAUCE));
+//				repo.save(new Ingredient("SRCR", "Sour Cream", Type.SAUCE));
+//			}
+//		};
+//	}
 }
