+++ src/main/java/axsoup/web/OrderController.java
@@ -5,7 +5,6 @@ import javax.validation.Valid;
 
 import axsoup.data.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.GetMapping;
