+++ src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,7 @@
+{"properties": [
+  {
+    "name": "soup.orders.page-size",
+    "type": "java.lang.String",
+    "description": "Sets the maximum number of orders to display in a list."
+  }
+]}
