+++ src/main/resources/application.properties
@@ -1,3 +1,27 @@
+#spring:
+#  security:
+#    user:
+#      name: buzz
+#      password: infinity
 
-hibernate.hbm2ddl.auto
-spring.jpa.show-sql=true
+#axsoup:
+#orders:
+#pageSize: 10
+#
+#
+#discount:
+#codes:
+#abcdef: 10
+#
+#---
+#spring:
+#profiles: prod
+#
+#datasource:
+#url: jdbc:mysql://localhost/soupcloud
+#username: tacouser
+#password: tacopassword
+#
+#logging:
+#level:
+#axsoup: WARN
